requirements.txt
langchain-community
langchain-core
langchain
langchain-openai
langchain-chroma
python-dotenv
bs4
ragas
datasets
pandas